aura-backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Cryptography & ZK
pynacl==1.5.0
cryptography==42.0.2
py_ecc==7.0.0
galois==0.3.8
numpy==1.26.4

# Cardano Integration
pycardano==0.10.0
blockfrost-python==0.6.0

# IPFS
ipfshttpclient==0.8.0a2
aiohttp==3.9.3

# AI/ML for Agents
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
factory-boy==3.3.0

# Development
python-dotenv==1.0.1
black==24.1.1
isort==5.13.2
mypy==1.8.0
